Remove broken UTF-8 chars in Japanese translation
The right solution is obviously to get someone to retranslate this line
and make sure it's correct, but for now this will allow the translation
to work again.

Mopy/bash/bash.py:
Noop change to make this a bit safer. Don't like using something dynamic
in safety= when we can use something static instead.